move the stuff around to prepare for merging the individual BTSs together
