+++ apps/trackers/constants.py
@@ -1,5 +1,5 @@
-from osidb.helpers import get_env
+"""
+common tracker constants
+"""
 
 TRACKER_API_VERSION = "v1"
-
-JIRA_SERVER = get_env("JIRA_URL", default="https://issues.redhat.com")
